Crate chain_impl_mockchain

Modules

account

accounting

block

Representation of the block in the mockchain.

certificate

chaineval

chaintypes

Shared Types between ledger and chain

config

date 🔒

error

Errors taht may happen in the ledger and mockchain.

evm

EVM transactions

fee

fragment

header

key

Module provides cryptographic utilities and types related to the user keys.

leadership

ledger

legacy

milli

multisig

multiverse

Multiverse

rewards

setting

define the Blockchain settings

stake

testing

tokens

transaction

treasury

update

utxo

Unspend Transaction Output (UTXO) ledger

value

vote

cover all that relate to the voting part of things (except for the certificate that are in the certificate module).