Type Alias chain_impl_mockchain::account::Ledger

pub type Ledger = Ledger<Identifier, ()>;

The public ledger of all accounts associated with their current state

Aliased Type

struct Ledger(Hamt<DefaultHasher, Identifier, AccountState<()>>);

Fields
0: Hamt<DefaultHasher, Identifier, AccountState<()>>
Implementations

impl<ID: Clone + Eq + Hash, Extra: Clone> Ledger<ID, Extra>

pub fn new() -> Self

Create a new empty account ledger

pub fn add_account( &self, identifier: ID, initial_value: Value, extra: Extra ) -> Result<Self, LedgerError>

Add a new account into this ledger.

If the identifier is already present, error out.

pub fn set_delegation( &self, identifier: &ID, delegation: &DelegationType ) -> Result<Self, LedgerError>

Set the delegation of an account in this ledger

pub fn exists(&self, identifier: &ID) -> bool

check if an account already exist

pub fn get_state( &self, account: &ID ) -> Result<&AccountState<Extra>, LedgerError>

Get account state

If the identifier does not match any account, error out

pub fn remove_account(&self, identifier: &ID) -> Result<Self, LedgerError>

Remove an account from this ledger

If the account still have value > 0, then error

pub fn add_value( &self, identifier: &ID, value: Value ) -> Result<Self, LedgerError>

Add value to an existing account.

If the account doesn’t exist, error out.

pub fn add_value_or_account( &self, identifier: &ID, value: Value, extra: Extra ) -> Result<Self, ValueError>

Add value to an existing account.

If the account doesn’t exist, it creates it with the value

pub fn add_rewards_to_account( &self, identifier: &ID, epoch: Epoch, value: Value, extra: Extra ) -> Result<Self, ValueError>

Add rewards to an existing account.

If the account doesn’t exist, it creates it with the value

pub fn spend( &self, identifier: &ID, counter: SpendingCounter, value: Value ) -> Result<Self, LedgerError>

Spend value from an existing account.

If the account doesn’t exist, or if the value is too much to spend, or if the spending counter doesn’t match, it throws a LedgerError.

pub(crate) fn spend_with_no_counter_check( &self, identifier: &ID, counter: SpendingCounter, value: Value ) -> Result<Self, LedgerError>

Spend value from an existing account without spending counter check.

If the account doesn’t exist, or if the value is too much to spend, it throws a LedgerError.

pub fn get_total_value(&self) -> Result<Value, ValueError>

pub fn token_add( &self, identifier: &ID, token: TokenIdentifier, value: Value ) -> Result<Self, LedgerError>

pub fn iter(&self) -> Iter<'_, ID, Extra>