Struct chain_impl_mockchain::accounting::account::account_state::AccountState

pub struct AccountState<Extra> {
    pub spending: SpendingCounterIncreasing,
    pub delegation: DelegationType,
    pub value: Value,
    pub tokens: Hamt<DefaultHasher, TokenIdentifier, Value>,
    pub last_rewards: LastRewards,
    pub extra: Extra,
}

Fields
spending: SpendingCounterIncreasingdelegation: DelegationTypevalue: Valuetokens: Hamt<DefaultHasher, TokenIdentifier, Value>last_rewards: LastRewardsextra: Extra
Implementations

impl<Extra> AccountState<Extra>

pub fn new(v: Value, e: Extra) -> Self

Create a new account state with a specific start value

pub fn new_reward(epoch: Epoch, v: Value, extra: Extra) -> Self

pub fn delegation(&self) -> &DelegationType

Get referencet to delegation setting

pub fn value(&self) -> Value

impl<Extra: Clone> AccountState<Extra>

pub fn add_value(&self, v: Value) -> Result<Self, ValueError>

Same as add() except use a ValueError

pub fn add(&self, v: Value) -> Result<Self, LedgerError>

Add a value to an account state

Only error if value is overflowing

pub fn add_rewards(&self, e: Epoch, v: Value) -> Result<Self, ValueError>

Add Rewards to the account value but also as the last_reward

pub fn spend( &self, spending: SpendingCounter, v: Value ) -> Result<Option<Self>, LedgerError>

Spends value from an account state, and returns the new state.

Note that this also increments the counter, as this function is usually used for spending.

pub(crate) fn spend_unchecked( &self, counter: SpendingCounter, v: Value ) -> Result<Option<Self>, LedgerError>

Spends value from an account state, and returns the new state.

Note that this also increments the counter, but does not fail if the given counter fails to match the current one. However, it does throw a warning.

pub fn token_add( &self, token: TokenIdentifier, v: Value ) -> Result<Self, LedgerError>

Add a value to a token in an account state

Only error if value is overflowing

pub fn set_delegation(&self, delegation: DelegationType) -> Self

Set delegation