Module chain_impl_mockchain::header

Re-exports

pub use crate::chaintypes::ChainLength;

pub use crate::chaintypes::HeaderId;

Modules

test

Structs

BftProof

BftSignature

BlockDate

Non unique identifier of the transaction position in the blockchain. There may be many transactions related to the same SlotId.

Common

GenesisPraosProof

HeaderBft

Finalized BFT Header

HeaderBftBuilder

Header Builder (finalize BFT part)

HeaderBuilder

Finalized BFT Header

HeaderDesc

Header description

HeaderGenesisPraos

Finalized Genesis-Praos Header

HeaderGenesisPraosBuilder

Header Builder (finalize Genesis-Praos part)

HeaderUnsigned

Finalized Unsigned Header

KesSignature

VrfProof

Enums

AnyBlockVersion

BlockVersion

Header

HeaderSetConsensusData

HeaderSetConsensusSignature

Proof

Functions

header_builder

Type Definitions

Epoch

HeaderBuilderNew

SlotId