Type Definition chain_impl_mockchain::header::builder::HeaderBuilderNew

pub type HeaderBuilderNew = HeaderBuilder<HeaderSetParenting>;

Implementations

impl HeaderBuilderNew

pub fn new(version: BlockVersion, contents: &Contents) -> Self

Create a new Header builder starting from the full content.

This doesn’t need the content directly, but only uses the content to calculate the content hash and the content size, and make sure this is consistent

pub fn new_raw( version: BlockVersion, content_hash: &BlockContentHash, content_size: BlockContentSize ) -> Self

recommended to use new(), this is only for test