Struct chain_impl_mockchain::leadership::Leadership

pub struct Leadership {
    epoch: Epoch,
    era: TimeEra,
    inner: LeadershipConsensus,
}

Leadership represent a given epoch and their associated leader or metadata.

Fields
epoch: Epochera: TimeErainner: LeadershipConsensus
Implementations

impl Leadership

pub fn new(epoch: Epoch, ledger: &Ledger) -> Self

pub fn epoch(&self) -> Epoch

get the epoch associated to the Leadership

pub fn stake_distribution(&self) -> Option<&StakeDistribution>

pub fn date_at_slot(&self, slot_id: u32) -> BlockDate

Create a Block date given a leadership and a relative epoch slot

Panics

If the slot index is not valid given the leadership, out of bound date

pub fn era(&self) -> &TimeEra

get the TimeEra associated to the Leadership

pub fn consensus(&self) -> &LeadershipConsensus

get the consensus associated with the Leadership

pub fn verify(&self, block_header: &Header) -> Verification

Verify whether this header has been produced by a leader that fits with the leadership

pub fn is_leader_for_date( &self, leader: &Leader, date: BlockDate ) -> LeaderOutput

Test that the given leader object is able to create a valid block for the leadership at a given date.