Module chain_impl_mockchain::ledger

Re-exports

pub use iter::*;

pub use ledger::*;

Modules

check

governance

define how to rule over the blockchain

iter

ledger

Mockchain ledger. Ledger exists in order to update the current state and verify transactions.

recovery

This module holds all the necessary methods to be able to serialize the ledger. There are 2 methods for each type that needs to be serialize pack_* and unpack_*.

token_distribution

Structs

EpochRewardsInfo

The epoch reward information.

LeadersParticipationRecord

Count how many blocks have been created by a specific Pool

Pots

Special pots of money

RewardsInfoParameters

Control what information will be extracted from the rewarding distribution process