Module chain_impl_mockchain::ledger::recovery

This module holds all the necessary methods to be able to serialize the ledger. There are 2 methods for each type that needs to be serialize pack_* and unpack_*.

The pack methods takes a mutable chain-core::packer::Codec<std::io::Write> reference and a reference to the type and writes the selected serialize format to the writer, it returns an std::io::Error wrapped in a Result if something goes wrong:

fn pack_<W: std::io::Write>(
    type: &T,
    codec: &mut Codec<W>,
) -> Result<(), std::io::Error> { ... }

The unpack method takes a mutable chain-core::packer::Codecstd::io::BufRead reference and returns an instance of a type wrapped in a Result

fn unpack_(
    codec: &mut Codec<&[u8]>,
) -> Result<T, std::io::Error> { ... }

For serializing the Ledger the approach is simple:

Iterate the Ledger

Pack each entry

Flag the end of packing

For deserializing:

Load all serialized Entry into a Vec

Use the from_iter ledger method to load it from the newly created vector.

Notice that the ledger::iter::Entry type holds references to the data types but when loading them from the serialized object we need to hold them. That is why we use the EntryOwned type instead for deserializing. This data is then cloned as necessary into the final deserialized ledger.

Enums

EntrySerializeCode ğŸ”’

Functions

pack_account_identifier ğŸ”’

pack_account_state ğŸ”’

pack_address ğŸ”’

pack_block_date ğŸ”’

pack_committee_public_keys ğŸ”’

pack_config_param ğŸ”’

pack_decl_element ğŸ”’

pack_declaration ğŸ”’

pack_delegation_ratio ğŸ”’

pack_delegation_type ğŸ”’

pack_digestof ğŸ”’

pack_discrimination ğŸ”’

pack_entry ğŸ”’

pack_globals ğŸ”’

pack_header_id ğŸ”’

pack_last_rewards ğŸ”’

pack_ledger_static_parameters ğŸ”’

pack_multisig_identifier ğŸ”’

pack_old_addr ğŸ”’

pack_output ğŸ”’

pack_payload_type ğŸ”’

pack_pool_id ğŸ”’

pack_pool_last_rewards ğŸ”’

pack_pool_registration ğŸ”’

pack_pool_state ğŸ”’

pack_pot_entry ğŸ”’

pack_spending_strategy ğŸ”’

pack_update_proposal ğŸ”’

pack_update_proposal_id ğŸ”’

pack_update_proposal_state ğŸ”’

pack_utxo_entry ğŸ”’

pack_vote_plan ğŸ”’

pack_vote_proposal ğŸ”’

pack_vote_proposals ğŸ”’

pack_voting_token ğŸ”’

unpack_account_identifier ğŸ”’

unpack_account_state ğŸ”’

unpack_address ğŸ”’

unpack_block_date ğŸ”’

unpack_committee_public_keys ğŸ”’

unpack_config_param ğŸ”’

unpack_decl_element ğŸ”’

unpack_declaration ğŸ”’

unpack_delegation_ratio ğŸ”’

unpack_delegation_type ğŸ”’

unpack_digestof ğŸ”’

unpack_discrimination ğŸ”’

unpack_entries ğŸ”’

unpack_entry_owned ğŸ”’

unpack_globals ğŸ”’

unpack_header_id ğŸ”’

unpack_last_rewards ğŸ”’

unpack_ledger_static_parameters ğŸ”’

unpack_multisig_identifier ğŸ”’

unpack_old_addr ğŸ”’

unpack_output ğŸ”’

unpack_payload_type ğŸ”’

unpack_pool_id ğŸ”’

unpack_pool_last_rewards ğŸ”’

unpack_pool_registration ğŸ”’

unpack_pool_state ğŸ”’

unpack_pot_entry ğŸ”’

unpack_proposal ğŸ”’

unpack_proposals ğŸ”’

unpack_spending_strategy ğŸ”’

unpack_update_proposal ğŸ”’

unpack_update_proposal_id ğŸ”’

unpack_update_proposal_state ğŸ”’

unpack_utxo_entry_owned ğŸ”’

unpack_vote_action ğŸ”’

unpack_vote_plan ğŸ”’

unpack_voting_token ğŸ”’