Function chain_impl_mockchain::rewards::rewards_contribution_calculation

pub fn rewards_contribution_calculation(
    epoch: Epoch,
    params: &Parameters,
    system_info: &SystemInformation
) -> Value

Calculate the reward contribution from the parameters

Note that the contribution in the system is still bounded by the remaining rewards pot, which is not taken in considering for this calculation.