Module chain_impl_mockchain::stake

Modules

controlled 🔒

delegation 🔒

distribution 🔒

stake 🔒

Structs

PercentStake

PoolLastRewards

PoolStakeDistribution

PoolStakeInformation

PoolState

PoolsState

A structure that keeps track of stake keys and stake pools.

SplitValueIn

Stake

StakeControl

StakeDistribution

Stake distribution at a given time.

StakeUnit

Enums

PoolError

Functions

get_distribution

Calculate the Stake Distribution where the source of stake is coming from utxos and accounts, and where the main targets is to calculate each value associated with known stake pools.