Struct chain_impl_mockchain::stake::controlled::StakeControl

pub struct StakeControl {
    assigned: Stake,
    unassigned: Stake,
    control: Hamt<DefaultHasher, Identifier, Stake>,
}

Fields
assigned: Stakeunassigned: Stakecontrol: Hamt<DefaultHasher, Identifier, Stake>
Implementations

impl StakeControl

pub fn new() -> Self

fn update_accounts(&self, accounts: &Ledger) -> Self

fn update_utxos(&self, utxos: &Ledger<Address>) -> Self

pub fn new_with(accounts: &Ledger, utxos: &Ledger<Address>) -> Self

pub fn total(&self) -> Stake

pub fn assigned(&self) -> Stake

pub fn unassigned(&self) -> Stake

pub fn by(&self, identifier: &Identifier) -> Option<Stake>

get the total stake controlled by the given account

pub fn ratio_by(&self, identifier: &Identifier) -> Ratio

get the ratio controlled by the given account

the ratio is based on the total assigned stake, stake that is not controlled (that is in UTxO without account keys) are not part of the equation.

pub fn add_unassigned(&self, stake: Stake) -> Self

pub fn remove_unassigned(&self, stake: Stake) -> Self

pub fn add_to(&self, identifier: Identifier, stake: Stake) -> Self

add the given amount of stake to the given identifier

also update the total stake

pub fn remove_from(&self, identifier: Identifier, stake: Stake) -> Self

add the given amount of stake to the given identifier

also update the total stake