Module chain_impl_mockchain::testing

Re-exports

pub use ledger::ConfigBuilder;

pub use ledger::LedgerBuilder;

pub use ledger::TestLedger;

pub use ledger::UtxoDb;

pub use arbitrary::*;

pub use builders::*;

Modules

arbitrary

builders

data

ledger

scenario

serialization

verifiers

Structs

KeysDb

TestGen

TestGen is a clone of abritrary architecture. There is a need to generate random struct, which would be used just once (quickcheck run test method multiple times thus generating time-consuming test case). This is needed for functional test approach rather than quickcheck approach

VoteTestGen