Module chain_impl_mockchain::testing::builders

Re-exports

pub use block_builder::*;

pub use cert_builder::*;

pub use old_address_builder::*;

pub use stake_pool_builder::*;

pub use tx_builder::*;

pub use witness_builder::*;

Modules

block_builder

cert_builder

initial_builder 🔒

old_address_builder

stake_pool_builder

tx_builder

tx_cert_builder 🔒

vote 🔒

witness_builder

Structs

FaultTolerantTxCertBuilder

this struct can create any transaction including not valid one in order to test robustness of ledger

InitialFaultTolerantTxBuilder

InitialFaultTolerantTxCertBuilder

TestTxCertBuilder

Enums

WitnessMode

Functions

create_initial_mint_token

create_initial_stake_pool_delegation

create_initial_stake_pool_owner_delegation

create_initial_stake_pool_registration

create_initial_stake_pool_update

Below method should be used only for negative testing

create_initial_transaction

create_initial_update_proposal

create_initial_update_vote

create_initial_vote_cast

create_initial_vote_plan

create_initial_vote_tally

decrypt_tally

evm_mapping_sign

plan_sign

pool_owner_sign

pool_owner_signed

tally_sign

update_proposal_sign

update_vote_sign