Module chain_impl_mockchain::testing::builders::witness_builder

Enums

Witness

Structure that proofs that certain user agrees with some data. This structure is used to sign Transaction and get SignedTransaction out.

Functions

make_witness

make_witness_with_lane

make_witnesses