Function chain_impl_mockchain::testing::serialization::serialization_bijection

pub fn serialization_bijection<T>(t: T) -> TestResultwhere
    T: Arbitrary + Serialize + DeserializeFromSlice + Eq,

test that any arbitrary given object can serialize and deserialize back into itself (i.e. it is a bijection, or a one to one match between the serialized bytes and the object)