Module chain_impl_mockchain::vote

cover all that relate to the voting part of things (except for the certificate that are in the certificate module).

Structs

Choice

a choice

CommitteeId

committee identifier

EncryptedVote

Options

options for the vote

ProofOfCorrectVote

TallyResult

the tally results

VotePlanLedger

VotePlanManager

Manage the vote plan and the associated votes in the ledger

VotePlanStatus

VoteProposalStatus

Weight

weight of a vote

Enums

Payload

PayloadType

the PayloadType to use for a vote plan

PrivateTallyState

Tally

TallyError

TryFromIntError

ValidatedPayload

VoteError

VotePlanLedgerError

Functions

encrypt_vote