Struct chain_impl_mockchain::vote::committee::CommitteeId

pub struct CommitteeId([u8; 32]);

committee identifier

this value is used to identify a committee member on chain as well as to use as input for the vote casting payload.

Tuple Fields
0: [u8; 32]
Implementations

impl CommitteeId

pub const COMMITTEE_ID_SIZE: usize = 32usize

pub fn to_hex(self) -> String

returns the identifier encoded in hexadecimal string

pub fn from_hex(s: &str) -> Result<Self, FromHexError>

read the identifier from the hexadecimal string

pub fn public_key(&self) -> PublicKey<Ed25519>