Enum chain_impl_mockchain::vote::manager::ProposalManagers

enum ProposalManagers {
    Public {
        managers: Vec<ProposalManager>,
    },
    Private {
        managers: Vec<ProposalManager>,
        crs: Arc<Crs>,
        election_pk: Arc<ElectionPublicKey>,
    },
}

Variants

Public

Fields

managers: Vec<ProposalManager>

Private

Fields

managers: Vec<ProposalManager>

crs: Arc<Crs>

election_pk: Arc<ElectionPublicKey>

Implementations

impl ProposalManagers

fn new(plan: &VotePlan) -> Self

fn managers(&self) -> &[ProposalManager]

fn managers_mut(&mut self) -> &mut [ProposalManager]

pub fn vote( &self, identifier: Identifier, vote_cast: ValidatedVoteCast, token_distribution: &TokenDistribution<'_, TokenIdentifier> ) -> Result<Self, VoteError>

Attempt to apply the vote to one of the proposals.

pub fn public_tally<F>( &self, token_distribution: &TokenDistribution<'_, TokenIdentifier>, governance: &Governance, f: F ) -> Result<Self, TallyError>where F: FnMut(&VoteAction),

pub fn validate_vote( &self, identifier: &Identifier, cast: VoteCast ) -> Result<ValidatedVoteCast, VoteError>

validate the vote against the proposal: verify that the proposal exists and the the length of the ciphertext is correct (if applicable)

pub fn finalize_private_tally<F>( &self, committee_pks: &[MemberPublicKey], decrypted_tally: &DecryptedPrivateTally, governance: &Governance, token_distribution: &TokenDistribution<'_, TokenIdentifier>, f: F ) -> Result<Self, VoteError>where F: FnMut(&VoteAction),