Struct chain_impl_mockchain::vote::VotePlanManager

pub struct VotePlanManager {
    id: VotePlanId,
    plan: Arc<VotePlan>,
    committee: Arc<HashSet<CommitteeId>>,
    proposal_managers: ProposalManagers,
}

Manage the vote plan and the associated votes in the ledger

this structure manage the lifespan of the vote plan, the votes casted and the associated parameters

Fields
id: VotePlanIdplan: Arc<VotePlan>committee: Arc<HashSet<CommitteeId>>proposal_managers: ProposalManagers
Implementations

impl VotePlanManager

pub fn new(plan: VotePlan, committee: HashSet<CommitteeId>) -> Self

pub fn id(&self) -> &VotePlanId

pub fn plan(&self) -> &VotePlan

pub fn statuses(&self) -> VotePlanStatus

pub fn can_vote(&self, date: BlockDate) -> bool

pub fn can_committee(&self, date: BlockDate) -> bool

pub fn committee_set(&self) -> &HashSet<CommitteeId>

pub fn vote_plan_elapsed(&self, date: BlockDate) -> bool

return true if the vote plan has elapsed i.e. the vote is no longer interesting to track in the ledger and it can be GCed.

fn valid_committee(&self, id: &CommitteeId) -> bool

pub fn vote( &self, block_date: BlockDate, identifier: Identifier, cast: VoteCast, token_distribution: TokenDistribution<'_, ()> ) -> Result<Self, VoteError>

attempt to apply the vote to one of the proposals

If the given identifier already had a vote, the previous vote will be discarded and only the new one will be kept.

errors

this function may fail if the proposal identifier is different

if the proposal index is not one one of the proposal listed

if the block_date show it is no longer valid to cast a vote for any of the managed proposals

if the payload type of the vote is not the expected one

pub fn public_tally<F>( &self, block_date: BlockDate, governance: &Governance, sig: CommitteeId, token_distribution: TokenDistribution<'_, ()>, f: F ) -> Result<Self, VoteError>where F: FnMut(&VoteAction),

pub fn private_tally<F>( &self, block_date: BlockDate, decrypted_tally: &DecryptedPrivateTally, governance: &Governance, sig: CommitteeId, token_distribution: TokenDistribution<'_, ()>, f: F ) -> Result<Self, VoteError>where F: FnMut(&VoteAction),