Module chain_impl_mockchain::certificate

Structs

DecryptedPrivateTally

DecryptedPrivateTallyError

DecryptedPrivateTallyProposal

EvmMapping

ExternalProposalDocument

abstract tag type to represent an external document, whatever it may be

MintToken

OwnerStakeDelegation

A self delegation to a specific StakePoolId.

PoolPermissions

Permission system related to the pool

PoolRegistration

Pool information

PoolRetirement

Retirement info for a pool

PoolUpdate

Updating info for a pool

Proposal

a proposal with the associated external proposal identifier which leads to retrieving data from outside of the blockchain with its unique identifier and the funding plan required for the proposal to be operated.

Proposals

a collection of proposals

StakeDelegation

UpdateProposal

UpdateVote

VoteCast

VotePlan

a vote plan for the voting system

VotePlanProof

VoteTally

Enums

Certificate

CertificatePayload

CertificateSlice

PoolSignature

PushProposal

SignedCertificate

TallyProof

VoteAction

this is the action that will result of the vote

VoteTallyPayload

Type Definitions

BftLeaderBindingSignature

ExternalProposalId

the identifier of the external proposal is the Blake2b 256 bits of the external proposal document hash.

GenesisPraosLeaderHash

Hash of keys used for pool

IndexSignatures

signatures with indices

ManagementThreshold

PoolId

Pool ID

PoolOwnersSigned

PoolRegistrationHash

Pool Registration Cryptographic Hash

UpdateProposalId

UpdateProposerId

UpdateVoterId

VotePlanId

the vote plan identifier on the blockchain