Struct chain_impl_mockchain::ledger::ledger::Ledger

pub struct Ledger { /* private fields */ }

Overall ledger structure.

This represent a given state related to utxo/old utxo/accounts/… at a given point in time.

The ledger can be easily and cheaply cloned despite containing reference to a lot of data (millions of utxos, thousands of accounts, ..)

Implementations

impl Ledger

pub fn stats(&self) -> Vec<String>

pub fn info_eq(&self, other: &Self) -> Vec<String>

impl Ledger

pub fn iter(&self) -> LedgerIterator<'_>

impl Ledger

pub fn new<'a, I>(
block0_initial_hash: HeaderId,
contents: I
) -> Result<Self, Error>where
I: IntoIterator<Item = &'a Fragment>,

pub fn can_distribute_reward(&self) -> bool

pub fn apply_protocol_changes(&self) -> Result<Self, Error>

pub fn distribute_rewards(
&self,
distribution: &StakeDistribution,
rewards_info_params: RewardsInfoParameters
) -> Result<(Self, EpochRewardsInfo), Error>

This need to be called before the first block of a new epoch

Reset the leaders log

Distribute the contribution (rewards + fees) to pools and their delegatees

pub fn begin_block(
&self,
chain_length: ChainLength,
block_date: BlockDate
) -> Result<ApplyBlockLedger, Error>

pub fn apply_block(
&self,
contents: &Contents,
metadata: &HeaderContentEvalContext
) -> Result<Self, Error>

Try to apply messages to a State, and return the new State if successful

pub fn apply_fragment(
&self,
content: &Fragment,
block_date: BlockDate
) -> Result<Self, Error>

Try to apply a message to the State, and return the new State if successful

this does not advance the state to the new state but apply a simple fragment of block to the current context.

pub fn apply_transaction<'a, Extra>(
self,
fragment_id: &FragmentId,
tx: &TransactionSlice<'a, Extra>
) -> Result<(Self, Value), Error>where
Extra: Payload,
LinearFee: FeeAlgorithm,

Applies transaction to the account ledger by validating and verifying inputs and outputs, as well as spending the fee value from the related account id. Returns an Error if the account cannot afford the fee or if the spending counter doesn’t match.

pub fn apply_update(self, update: &UpdateProposal) -> Result<Self, Error>

pub fn apply_update_proposal(
self,
proposal_id: UpdateProposalId,
proposal: UpdateProposal,
auth_data: &TransactionBindingAuthData<'_>,
sig: BftLeaderBindingSignature,
cur_date: BlockDate
) -> Result<Self, Error>

pub fn apply_update_vote<'a>(
self,
vote: &UpdateVote,
auth_data: &TransactionBindingAuthData<'a>,
sig: BftLeaderBindingSignature
) -> Result<Self, Error>

pub fn apply_vote_plan(
self,
tx: &TransactionSlice<'_, VotePlan>,
cur_date: BlockDate,
vote_plan: VotePlan,
sig: VotePlanProof
) -> Result<Self, Error>

pub fn apply_vote_cast(
self,
account_id: Identifier,
vote: VoteCast
) -> Result<Self, Error>

pub fn active_vote_plans(&self) -> Vec<VotePlanStatus>

pub fn apply_vote_tally<'a>(
self,
tally: &VoteTally,
bad: &TransactionBindingAuthData<'a>,
sig: TallyProof
) -> Result<Self, Error>

pub fn apply_pool_registration_signcheck<'a>(
self,
cert: &PoolRegistration,
bad: &TransactionBindingAuthData<'a>,
sig: PoolSignature
) -> Result<Self, Error>

pub fn apply_pool_registration(
self,
cert: &PoolRegistration
) -> Result<Self, Error>

pub fn apply_pool_retirement<'a>(
self,
auth_cert: &PoolRetirement,
bad: &TransactionBindingAuthData<'a>,
sig: PoolSignature
) -> Result<Self, Error>

pub fn apply_pool_update<'a>(
self,
auth_cert: &PoolUpdate,
bad: &TransactionBindingAuthData<'a>,
sig: PoolSignature
) -> Result<Self, Error>

pub fn apply_stake_delegation(
self,
auth_cert: &StakeDelegation
) -> Result<Self, Error>

pub fn apply_owner_stake_delegation(
self,
account_id: &UnspecifiedAccountIdentifier,
witness: &Witness,
delegation: &OwnerStakeDelegation
) -> Result<Self, Error>

pub fn mint_token(self, mt: MintToken) -> Result<Self, Error>

pub fn mint_token_unchecked(self, mt: MintToken) -> Result<Self, Error>

pub fn get_stake_distribution(&self) -> StakeDistribution

pub fn get_static_parameters(&self) -> &LedgerStaticParameters

access the ledger static parameters

pub fn accounts(&self) -> &Ledger

pub fn updates(&self) -> &UpdateState

pub fn token_totals(&self) -> &TokenTotals

pub fn token_distribution(&self) -> TokenDistribution<'_, ()>

pub fn consensus_version(&self) -> ConsensusType

pub fn utxo_out(
&self,
fragment_id: FragmentId,
index: TransactionIndex
) -> Option<&Output<Address>>

pub fn utxos(&self) -> Iter<'_, Address>

pub fn chain_length(&self) -> ChainLength

pub fn settings(&self) -> &Settings

pub fn delegation(&self) -> &PoolsState

pub fn delegation_mut(&mut self) -> &mut PoolsState

pub fn date(&self) -> BlockDate

pub fn era(&self) -> &TimeEra

pub fn get_total_value(&self) -> Result<Value, Error>

pub fn remaining_rewards(&self) -> Value

pub fn treasury_value(&self) -> Value