Module chain_impl_mockchain::transaction

Modules

test

Structs

Input

Generalized input which have a specific input value, and either contains an account reference or a TransactionSignDataHash+index

InputOutput

Inputs & Outputs for a built transaction

InputOutputBuilder

Inputs & Outputs for a transaction being built

InputsIter

InputsSlice

InputsWitnessesIter

InputsWitnessesSlice

NoExtra

Output

Information how tokens are spent. A value of tokens is sent to the address.

OutputsIter

OutputsSlice

PayloadAuthData

Owned binary representation of a payload auth

PayloadAuthSlice

Borrowed binary representation of a payload auth

PayloadData

Owned binary representation of a payload

PayloadSlice

Borrowed binary representation of a payload

SetAuthData

SetIOs

SetTtl

SetWitnesses

SingleAccountBindingSignature

Transaction

TransactionAuthData

TransactionBindingAuthData

TransactionBindingAuthDataPhantom

TransactionSignData

TransactionSlice

TxBuilderState

A Transaction builder with an associated state machine

UnspecifiedAccountIdentifier

This is either an single account or a multisig account depending on the witness type

UnverifiedTransactionSlice

UtxoPointer

Unspent transaction pointer.

WitnessAccountData

WitnessMultisigData

WitnessUtxoData

WitnessesIter

WitnessesSlice

Enums

AccountBindingSignature

AccountIdentifier

Balance

Amount of the balance in the transaction.

BalanceError

Error

Possible error for the builder.

InputEnum

InputType

OutputPolicy

Output policy to be used in transaction. This policy is used then there is positive balance on in the OUTPUT+FEE-INPUT. Policy explains how to use that balance. Rember that policy application may change the amount of the fee.

SetPayload

TransactionStructError

Witness

Structure that proofs that certain user agrees with some data. This structure is used to sign Transaction and get SignedTransaction out.

WitnessUtxoVersion

Constants

INPUT_PTR_SIZE

INPUT_SIZE

Traits

Payload

Type Definitions

AuthenticatedTransaction

TransactionIndex

TransactionSignDataHash

TxBuilder